Handling Lifecycles with Lifecycle-Aware Components
Lifecycle-aware components perform actions in response to a change in the lifecycle status of another component, such as activities and fragments.
These components help you produce better-organized,and often lighter-weight code, that is easier to maintain.

A common pattern is to implement the actions of the dependent components in the lifecycle methods of activities and fragments. However,
this pattern leads to a poor organization of the code and to the proliferation of errors.
By using lifecycle-aware components,you can move the code of dependent components out of the lifecycle methods and into the components themselves.

The androidx.lifecycle package provides classes and interfaces that let you build lifecycle-aware components—which are components that can automatically adjust their behavior based on the current lifecycle state of an activity or fragment.

There are two types of lifecycle aware components:
1) Lifecycle Owner
2) Lifecycle Observer

	1)Lifecycle owner are activities, fragment.
	2)Lifecycle Observer will observe the state of activities and fragments.